More work work work

+++ src/main/java/de/tum/in/www1/artemis/service/listeners/ResultListener.java
@@ -4,8 +4,7 @@ import javax.persistence.PostPersist;
 import javax.persistence.PostUpdate;
 import javax.persistence.PreRemove;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 
@@ -15,7 +14,8 @@ import de.tum.in.www1.artemis.service.ScoreService;
 @Component
 public class ResultListener {
 
-    private final Logger log = LoggerFactory.getLogger(ResultListener.class);
+    @Value("#{new Boolean('${artemis.resultListener.isActive}')}")
+    private Boolean isActive;
 
     private ScoreService scoreService;
 
@@ -37,6 +37,10 @@ public class ResultListener {
      */
     @PreRemove
     public void removeOrUpdateAssociatedParticipantScore(Result resultToBeDeleted) {
+        if (isActive == null || !isActive) {
+            return;
+        }
+
         scoreService.removeOrUpdateAssociatedParticipantScore(resultToBeDeleted);
     }
 
@@ -48,6 +52,10 @@ public class ResultListener {
     @PostUpdate
     @PostPersist
     public void updateOrCreateParticipantScore(Result result) {
+        if (isActive == null || !isActive) {
+            return;
+        }
+
         scoreService.updateOrCreateParticipantScore(result);
 
     }
